[MIFOS-3174] Several bug fixes around additional fields data migration

+++ application/src/main/java/org/mifos/accounts/loan/persistance/LoanDao.java
@@ -37,5 +37,5 @@ public interface LoanDao {
 
     List<CustomFieldDefinitionEntity> retrieveCustomFieldEntitiesForLoan();
 
-    List<AccountCustomFieldEntity> getCustomFieldResponses(Integer customFieldId);
+    List<AccountCustomFieldEntity> getCustomFieldResponses(Short customFieldId);
 }
